Simplify delta decoding
I just wrote a delta encoding/decoding crate that takes care of some edge cases and has a lot of testing and perf benchmarking, and this PR makes use of that crate.

I'm still hacking on the PBF encoding, so the delta encoding from that lib will be used as well.

Note that I use `copied()` iterator which simplifies value handling without adding any perf costs (dealing with values vs refs is simpler too).

The `WayRefIter` is technically not needed other than to implement the `ExactSizeIterator` on it.  If the exact size hint is not required, that whole struct can be replaced with this for backward compatibility.

```
pub type WayRefIter<'a> = DeltaDecoderIter<Copied<SliceIter<'a, i64>>>;
```